+++ src/lib/wapi/functions/delete-messages-me.js
@@ -52,47 +52,28 @@ MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMNMMNMNMMMNMMNNMMMMMMMMMMMM
 MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMNNNNMMNNNMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 */
-import { ConnectionTransport } from 'puppeteer';
 
-import * as WebSocket from 'ws';
+export async function deleteMessagesMe(chatId, messageArray, revoke) {
+  let userId = new window.Store.UserConstructor(chatId, {
+    intentionallyUsePrivateConstructor: true
+  });
+  let conversation = window.Store.Chat.get(chatId);
 
-export class WebSocketTransport implements ConnectionTransport {
-  static create(url: string, timeout?: number): Promise<WebSocketTransport> {
-    return new Promise((resolve, reject) => {
-      const ws = new WebSocket(url, [], {
-        perMessageDeflate: false,
-        maxPayload: 256 * 1024 * 1024, // 256Mb
-        handshakeTimeout: timeout
-      });
+  if (!conversation) return false;
 
-      ws.addEventListener('open', () => resolve(new WebSocketTransport(ws)));
-      ws.addEventListener('error', reject);
-    });
+  if (!Array.isArray(messageArray)) {
+    messageArray = [messageArray];
   }
 
-  private _ws: WebSocket;
-  onmessage?: (message: string) => void;
-  onclose?: () => void;
+  let messagesToDelete = messageArray.map((msgId) =>
+    window.Store.Msg.get(msgId)
+  );
 
-  constructor(ws: WebSocket) {
-    this._ws = ws;
-    this._ws.addEventListener('message', (event) => {
-      if (this.onmessage) this.onmessage.call(null, event.data);
-    });
-    this._ws.addEventListener('close', () => {
-      if (this.onclose) this.onclose.call(null);
-    });
-    // Silently ignore all errors - we don't know what to do with them.
-    this._ws.addEventListener('error', () => {});
-    this.onmessage = null;
-    this.onclose = null;
+  if (revoke) {
+    conversation.sendRevokeMsgs(messagesToDelete, conversation);
+  } else {
+    conversation.sendDeleteMsgs(messagesToDelete, conversation);
   }
 
-  send(message: string): void {
-    this._ws.send(message);
-  }
-
-  close(): void {
-    this._ws.close();
-  }
+  return true;
 }