+++ src/lib/wapi/functions/logout.js
@@ -53,8 +53,8 @@ MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMNNNNMMNNNMMMMMMMMMMMMMMMMM
 MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 */
 export async function logout() {
-  if (window.Store.ws2) {
-    await window.Store.ws2.logout();
+  if (window.Store.Login) {
+    await window.Store.Login.triggerCriticalSyncLogout();
     return true;
   } else {
     return false;