+++ src/lib/wapi/functions/send-link-preview.js
@@ -74,21 +74,24 @@ export async function sendLinkPreview(chatId, url, text) {
     }
   };
   if (!_Path.Reg().test(url)) {
-    var text =
-      'Use a valid HTTP protocol. Example: https://www.youtube.com/watch?v=V1bFr2SWP1';
-    return WAPI.scope(chatId, true, null, text);
+    return WAPI.scope(
+      chatId,
+      true,
+      null,
+      'Use a valid HTTP protocol. Example: https://www.youtube.com/watch?v=atPaQtpx5QQ'
+    );
   }
   var chat = await WAPI.sendExist(chatId);
   if (!chat.erro) {
     const linkPreview = await Store.WapQuery.queryLinkPreview(url);
-    const newMsgId = await window.WAPI.getNewMessageId(chat.id);
+    const newMsgId = await window.WAPI.getNewMessageId(chat.id._serialized);
     let inChat = await WAPI.getchatId(chat.id).catch(() => {});
     if (inChat) {
       chat.lastReceivedKey._serialized = inChat._serialized;
       chat.lastReceivedKey.id = inChat.id;
     }
 
-    const fromwWid = await window.Store.Conn.wid;
+    const fromwWid = await Store.MaybeMeUser.getMaybeMeUser();
     const message = {
       id: newMsgId,
       ack: 0,
@@ -101,13 +104,21 @@ export async function sendLinkPreview(chatId, url, text) {
       isNewMsg: !0,
       type: 'chat',
       subtype: 'url',
-      canonicalUrl: linkPreview.canonicalUrl,
-      description: linkPreview.description,
-      doNotPlayInline: linkPreview.doNotPlayInline,
-      matchedText: linkPreview.matchedText,
-      preview: linkPreview.preview,
-      thumbnail: linkPreview.thumbnail,
-      title: linkPreview.title
+      canonicalUrl: linkPreview.canonicalUrl
+        ? linkPreview.canonicalUrl
+        : undefined,
+      description: linkPreview.description
+        ? linkPreview.description
+        : undefined,
+      doNotPlayInline: linkPreview.doNotPlayInline
+        ? linkPreview.doNotPlayInline
+        : undefined,
+      matchedText: linkPreview.matchedText
+        ? linkPreview.matchedText
+        : undefined,
+      preview: linkPreview.preview ? linkPreview.preview : undefined,
+      thumbnail: linkPreview.thumbnail ? linkPreview.thumbnail : undefined,
+      title: linkPreview.title ? linkPreview.title : undefined
     };
     var result = (
       await Promise.all(window.Store.addAndSendMsgToChat(chat, message))