+++ src/lib/wapi/functions/set-profile-pic.js
@@ -1,7 +1,11 @@
 export async function setProfilePic(obj, id) {
   if (!id) {
-    id = Store.Me.attributes.wid._serialized;
+    id = await Store.MaybeMeUser.getMaybeMeUser();
   }
   let base64 = 'data:image/jpeg;base64,';
-  return await Store.Profile.sendSetPicture(id, base64 + obj.b, base64 + obj.a);
+  return await Store.Profile.sendSetPicture(
+    id._serialized,
+    base64 + obj.b,
+    base64 + obj.a
+  );
 }